requirements.txt
flask
supabase
python-dotenv  # <-- 방금 추가한 것
waitress       # 윈도우 로컬 실행용 (선택)
gunicorn       # Render(리눅스) 실행용